New data models, user sidebar, general cleanup.
New data models are in place to make sure we are aligned with the
backend project.  Map data is now separate from user data, the two are
combined on the fly at application start-up.  A new userData,json file
was added as a fill-in for some sample data from the backend.

There is a new user sidebar area, this will possibly replace the table
in the footer.  It includes a default profile picture for the moment.

General cleanup for some of the Javascript and HTML.